Use ASL template in more places (fixes #188, #187)
Use the 'asl_provided' variable and the corresponding template text in
more places.  After this change, every event post that contained one
of a) the 'asl_provided' variable in its front matter, or b) the
sentence "This event will have an American Sign Language interpreter"
or c) "This event will not have an American Sign Language interpreter"
now uses a standard template that is conditional on the variable.

In cases where the old sentence and variable disagreed, the variable
stays the same (and thus the sentence as displayed is now different,
once the template is processed).  In cases where the variable was
present but there was no sentence, the standard template sentence is
now present.

Posts that had neither the variable nor the sentence are unchanged;
that's mostly posts from 2016 and earlier.